requirements.txt
face_recognition==1.3.0
Flask==2.0.3
h5py==2.10.0
matplotlib==3.3.4
numpy==1.18.5
opencv_python==4.6.0.66
pandas==1.1.5
Pillow==9.1.1
scikit_learn==1.1.1
tensorflow==2.9.1
tensorflow_directml==1.15.7
